+++ src/cast.rs
@@ -19,7 +19,6 @@ use crate::{
     FixedU64, FixedU8,
 };
 use az_crate::{Cast, CheckedCast, OverflowingCast, SaturatingCast, UnwrappedCast, WrappingCast};
-#[cfg(feature = "f16")]
 use half::{bf16, f16};
 
 macro_rules! cast {
@@ -183,12 +182,7 @@ macro_rules! cast_num {
             $Fixed($LeEqU);
             i8, i16, i32, i64, i128, isize,
             u8, u16, u32, u64, u128, usize,
-            f32, f64, F128Bits,
-        }
-        #[cfg(feature = "f16")]
-        cast_num! {
-            $Fixed($LeEqU);
-            f16, bf16,
+            f16, bf16, f32, f64, F128Bits,
         }
     )* };
 }